TASK-13-1-2 Solution to WebRadio and HbbTV
The WebRadio-controller features active categories 'popular', 'recent' and
'all' as well as a filter acting on the station name.

Unfortunately the filter is not super-interactive, as it will only filter one
page as provided via the API. It would be nice to either enable an API-search
mechanism, or to fetch results in bulk from the API and perform
pagination/filtering locally.